Fix bug in client registration.
When performing batchSave for the client and client-history records there would be intermittent failures.
Since both record types use the client name as the partition key and a created-at timestamp as the range key, then it's possible that the timestamp is generated in the same millisecond.  The batchSave method will fail if provided with duplicate partition-key/range-key pairs.

Exception: Provided list of item keys contains duplicates